+++ 6020_Glaphic2.vcxproj
@@ -144,6 +144,7 @@
     <ClCompile Include="dev\inc\imgui\imgui_impl_opengl3.cpp" />
     <ClCompile Include="dev\inc\imgui\imgui_tables.cpp" />
     <ClCompile Include="dev\inc\imgui\imgui_widgets.cpp" />
+    <ClCompile Include="src\callBack.cpp" />
     <ClCompile Include="src\drawOBJ.cpp" />
     <ClCompile Include="src\FBO\cFBO.cpp" />
     <ClCompile Include="src\glad.c" />
