+++ 6020_Glaphic2.vcxproj.filters
@@ -117,6 +117,9 @@
     <ClCompile Include="src\lightSetup.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="src\callBack.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\GUI\cGUI.h">
